+++ config.py
@@ -3,15 +3,13 @@ import os
 
 CONFIG_FILE_PATH = "config.json"
 
-def create_default_config():
-    default_config = {
-        'browser': 'all',
-        'schedule_name': 'Hoyolab Check In',
-        'delay_minute': 1,
-        'run_time(24h)': 12,
-        'random_delay' : 5
-    }
-    return default_config
+default_config = {
+    'browser': 'all',
+    'schedule_name': 'Hoyolab Check In',
+    'delay_minute': 1,
+    'run_time(24h)': 12,
+    'random_delay' : 5
+}
 
 def write_config(config):
     with open(CONFIG_FILE_PATH, "w") as config_file:
@@ -19,17 +17,15 @@ def write_config(config):
 
 def read_config():
     if not os.path.exists(CONFIG_FILE_PATH):
-        config = create_default_config()
+        config = default_config
         write_config(config)
     else:
         with open(CONFIG_FILE_PATH, "r") as config_file:
             config = json.load(config_file)
-        default_config = create_default_config()
         for key, value in default_config.items():
             if key not in config:
                 config[key] = value
         write_config(config)
-
     return config
 
 def update_config(key, value):
